'downgrade' code to make it ES5 compatible fix ESlint errors
