+++ circlewrap.js
@@ -5,63 +5,48 @@
  * Helper class, ported from CirclePackLayout gephi plugin
  */
 
-class CircleWrap {
-    wrappedCircle;//hacky d3 reference thing
+function CircleWrap(id, x, y, r, circleWrap) {
+    this.wrappedCircle = circleWrap || null;//hacky d3 reference thing
 
-    children = new Map();
-    id;
-    next;
-    previous;
+    this.children = {};
+    this.id = id || null;
+    this.next = null;
+    this.previous = null;
 
-    x;
-    y;
-    r = 999;
+    this.x = x || null;
+    this.y = y || null;
+    if (circleWrap)
+        this.r = 1010101; // XXX should not be used in this case
+    else
+        this.r = r || 999;
 
-
-    constructor() {
-    }
-
-    static fromCircleWrap(cw) {
-        let ret = new CircleWrap();
-        ret.wrappedCircle = cw;
-        ret.r = 101010; // XXX
-        return ret;
-    }
-
-    static fromValues(x, y, r) {
-        if (r < 0)
-            throw new Error('graphology-layout/circlewrap: r cannot be < 0.');
-
-        let ret = new CircleWrap();
-        ret.x = x;
-        ret.y = y;
-        ret.r = r;
-        return ret;
-    }
-
-    hasChildren() {
-        return this.children.size > 0;
-    }
-    addChild(id, child) {
-        this.children.set(id, child);
-    }
-
-    getChild(id){
-        if (!this.children.has(id)){
-            let circleWrap = new CircleWrap();
-            this.children.set(id, circleWrap);
-        }
-        return this.children.get(id);
-    }
-
-    applyPositionToChildren() {
-        if (this.hasChildren()) {
-            this.children.forEach((child, key) => {
-               child.x += this.x;
-               child.y += this.y;
-               child.applyPositionToChildren();
-            });
-        }
-    }
 }
-module.exports = CircleWrap
+
+CircleWrap.prototype.hasChildren = function() {
+    return Object.keys(this.children).length > 0;
+};
+
+CircleWrap.prototype.addChild = function(id, child) {
+    this.children[id] = child;
+};
+
+CircleWrap.prototype.getChild = function(id) {
+    if (!this.children.hasOwnProperty(id)) {
+        var circleWrap = new CircleWrap();
+        this.children[id] = circleWrap;
+    }
+    return this.children[id];
+};
+
+CircleWrap.prototype.applyPositionToChildren = function() {
+    if (this.hasChildren()) {
+        Object.keys(this.children).forEach(function(key) {
+            var child = this.children[key];
+            child.x += this.x;
+            child.y += this.y;
+            child.applyPositionToChildren();
+        });
+    }
+};
+
+module.exports = CircleWrap;
